+++ subsys/bluetooth/controller/ll_sw/ull_peripheral_iso_internal.h
@@ -15,6 +15,7 @@ uint8_t ull_peripheral_iso_acquire(struct ll_conn *acl,
 uint8_t ull_peripheral_iso_setup(struct pdu_data_llctrl_cis_ind *ind,
 				 uint8_t cig_id,
 				 uint16_t cis_handle);
-void ull_peripheral_iso_start(struct ll_conn *acl, uint32_t ticks_at_expire,
-			      uint16_t cis_handle);
 void ull_peripheral_iso_update_peer_sca(struct ll_conn *acl);
+void ull_peripheral_iso_update_ticker(struct ll_conn_iso_group *cig,
+				      uint32_t ticks_at_expire,
+				      uint32_t iso_interval_us_frac);
